btrd: fix the handling of btrfs_shared_data_ref
The definition for `btrfs_shared_data_ref` was incorrect, it's actually
defined as

```
struct btrfs_shared_data_ref {
        __le32 count;
} __attribute__ ((__packed__));
```

In addition to that, we are only keeping the `type` portion of
`btrfs_extent_inline_ref` defined, but in the case of
`btrfs_shared_data_ref` we still have the `offset` field used, so we
need to add the size of u64 to offset to make sure we advance the right
amount in the leaf.

This fixes the error

```
Error: Unknown btrfs_extent_inline_ref type: 0
```

When trying to search the extent tree on a file system that has had a
snapshot taken and then the original file deleted from the origin
snapshot.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>